/  --iVersion iPhone Client ***
Refactored code so each sensor gets a draw and a print function.  This should make for smoother and quicker animation. 

It is an update the reading as it comes in strategy.  Whatever sensor comes over the wire is the only one that gets updated. Hopefully the timing is such that none of the sensors gets skipped.